// realteeth/src/shared/ui/modal/modal.type.ts
export type ModalVariants = 'default' | 'fullscreen';
export type ModalSizes = 'sm' | 'md' | 'lg' | 'xl';
export type ModalPositions = 'center' | 'top' | 'bottom';
export type ModalAnimationTypes = 'fade' | 'slide' | 'zoom';
export type ModalCloseButtonPositions =
  | 'top-right'
  | 'top-left'
  | 'bottom-right'
  | 'bottom-left'
  | 'none';